refactored out model loader code, sprite loader simplifyed
